+++ PaperMalKing.Shikimori.Wrapper/Models/List/BaseListEntry.cs
@@ -6,7 +6,7 @@ namespace PaperMalKing.Shikimori.Wrapper.Models.List
 	internal abstract class BaseListEntry<T> where T : BaseListSubEntry
 	{
 		[JsonPropertyName("id")]
-		public int Id { get; init; }
+		public ulong Id { get; init; }
 
 		[JsonPropertyName("score")]
 		public byte Score { get; init; }
