+++ PaperMalKing.Shikimori.Wrapper/Models/List/BaseListSubEntry.cs
@@ -9,7 +9,7 @@ namespace PaperMalKing.Shikimori.Wrapper.Models.List
 		public abstract string TotalAmount { get; }
 
 		[JsonPropertyName("id")]
-		public int Id { get; init; }
+		public ulong Id { get; init; }
 
 		[JsonPropertyName("name")]
 		public string Name { get; init; } = null!;
